+++ test-suite/testkit/src/api.rs
@@ -32,7 +32,7 @@ use std::{
 };
 
 use exonum::{
-    api::{self, node::public::system::DispatcherInfo, ApiAggregator},
+    api::{self, ApiAggregator},
     messages::{AnyTx, Verified},
     node::ApiSender,
 };
@@ -46,7 +46,8 @@ use crate::TestKit;
 /// [`TestKitApi`]: struct.TestKitApi.html
 #[derive(Debug, Clone, Copy)]
 pub enum ApiKind {
-    /// `api/system` endpoints of the built-in Exonum REST API.
+    /// `api/system` endpoints of the system API node plugin. To access endpoints, the plugin
+    /// should be attached to the testkit.
     System,
     /// Endpoints of the REST API of the explorer service.
     Explorer,
@@ -135,11 +136,6 @@ impl TestKitApi {
             ApiAccess::Private,
             kind.to_string(),
         )
-    }
-
-    /// Performs a GET request to the "/services" system endpoint.
-    pub fn dispatcher_info(&self) -> DispatcherInfo {
-        self.public(ApiKind::System).get("v1/services").unwrap()
     }
 }